# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DeliveryStatus": {
        "enum": [
          "queued",
          "accepted_by_gmail",
          "failed"
        ],
        "title": "DeliveryStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InviteMemberResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "InvitationPageResponse",
        "type": "object"
      },
      "InvitationPreview": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "invited_by_name": {
            "title": "Invited By Name",
            "type": "string"
          },
          "invited_email": {
            "title": "Invited Email",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvitationStatus"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "team_name",
          "invited_email",
          "role",
          "invited_by_name",
          "expires_at",
          "status"
        ],
        "title": "InvitationPreview",
        "type": "object"
      },
      "InvitationStatus": {
        "enum": [
          "pending",
          "accepted",
          "revoked",
          "expired"
        ],
        "title": "InvitationStatus",
        "type": "string"
      },
      "InviteMemberRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "role": {
            "const": "member",
            "default": "member",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "InviteMemberRequest",
        "type": "object"
      },
      "InviteMemberResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_attempts": {
            "title": "Delivery Attempts",
            "type": "integer"
          },
          "delivery_status": {
            "$ref": "#/components/schemas/DeliveryStatus"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvitationStatus"
          },
          "team_id": {
            "format": "uuid",
            "title": "Team Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "team_id",
          "email",
          "role",
          "status",
          "delivery_status",
          "delivery_attempts",
          "created_at",
          "expires_at"
        ],
        "title": "InviteMemberResponse",
        "type": "object"
      },
...
      "TeamMembershipResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "joined_at": {
            "format": "date-time",
            "title": "Joined At",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "team_id": {
            "format": "uuid",
            "title": "Team Id",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "team_id",
          "user_id",
          "role",
          "joined_at"
        ],
        "title": "TeamMembershipResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "VirtuJudge Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/invitations/{token}": {
      "get": {
        "operationId": "invitation_preview_api_v1_invitations__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invitation Preview",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{token}/accept": {
      "post": {
        "operationId": "accept_invitation_api_v1_invitations__token__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMembershipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/api/v1/teams/{team_id}/invitations": {
      "get": {
        "operationId": "list_team_invitations_api_v1_teams__team_id__invitations_get",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Team Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Team Invitations",
        "tags": [
          "teams"
        ]
      },
      "post": {
        "operationId": "invite_team_member_api_v1_teams__team_id__invitations_post",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Team Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteMemberResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Entity tag for optimistic concurrency control",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invite Team Member",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/invitations/{id}": {
      "delete": {
        "operationId": "revoke_team_invitation_api_v1_teams__team_id__invitations__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Team Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Team Invitation",
        "tags": [
          "teams"
        ]
      }
    },
...
  }
}